pyproject.toml
[project]
name = "sap-order-to-cash-dataset"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "groq>=1.1.1",
    "networkx>=3.6.1",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.2",
    "pyvis>=0.3.2",
    "streamlit>=1.55.0",
]
